fix(feishu): extract media chunk-idle helper under LOC ratchet
